factor out `hashMessages`
